extract postgres functions out to postgres_utils.hpp
